:sparkles: [feat] URI 내 query string parsing
